Reorganizing entity classes
Move cart related classes to Entity\Cart
Move order items classes to Entity\OrderElement
