+++ src/Demo/CartBundle/Entity/Cart/WishListInterface.php
@@ -6,7 +6,7 @@
  * Time: 4:51 PM
  */
 
-namespace Demo\CartBundle\Entity;
+namespace Demo\CartBundle\Entity\Cart;
 
 
 interface WishListInterface extends CartInterface
