+++ src/Demo/CartBundle/Entity/OrderElement/WishOrderItem.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Demo\CartBundle\Entity;
+namespace Demo\CartBundle\Entity\OrderElement;
 
 use Doctrine\ORM\Mapping as ORM;
